remove parachain dependency on tee
