+++ Cargo.toml
@@ -18,7 +18,9 @@ members = [
     'precompiles/*',
     'primitives/common',
     'primitives/core',
+    'primitives/core/proc-macros',
     'primitives/core/macros',
+    'primitives/hex',
     'primitives/sidechain',
     'primitives/teeracle',
     'primitives/teerex',
