+++ primitives/core/Cargo.toml
@@ -19,8 +19,9 @@ sp-io = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0
 sp-runtime = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.42", default-features = false }
 sp-std = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.42", default-features = false }
 
-itp-utils = { path = "../../tee-worker/core-primitives/utils", default-features = false }
+litentry-hex-utils = { path = "../hex", default-features = false }
 litentry-macros = { path = "macros" }
+litentry-proc-macros = { path = "proc-macros" }
 
 [features]
 default = ["std"]
@@ -35,6 +36,5 @@ std = [
     "sp-std/std",
     "sp-io/std",
     "ring/std",
-    "itp-utils/std",
     "pallet-evm/std",
 ]
