+++ primitives/core/macros/Cargo.toml
@@ -4,11 +4,3 @@ description = 'Proc-macros used by Litentry crates.'
 name = "litentry-macros"
 version = "0.9.12"
 edition = "2021"
-
-[lib]
-proc-macro = true
-
-[dependencies]
-proc-macro2 = { version = "1" }
-quote = { version = "1" }
-syn = { version = "2", features = ["full", "visit-mut"] }
