+++ primitives/core/macros/src/lib.rs
@@ -13,81 +13,24 @@
 //
 // You should have received a copy of the GNU General Public License
 // along with Litentry.  If not, see <https://www.gnu.org/licenses/>.
+#![cfg_attr(not(feature = "std"), no_std)]
 
-#![allow(clippy::tabs_in_doc_comments)]
-
-use proc_macro::TokenStream;
-use reuse::handle_reuse;
-use syn::{parse_macro_input, Error};
-
-mod reuse;
-
-/**
-This macro is used to reuse implementations when the rust's trait system cannot gracefully express the abstraction.
-
-This works similar with `#[cfg(..)]` that sets the target only appear on the specified cases.
-
-# Usage:
-
-```
-use litentry_macros::reuse;
-
-#[reuse(x, y)] . // Define the cases that the following implementation expands for each one
-mod __ {  // Leave mod name with double discards, which is to be replaced by the cases
-	#[x]  // This item would only appear on case `x`
-	fn u() {
-		__
-	}
-
-	#[y]  // This item would only appear on case `y`
-	fn v(a: String) {
-		println!("hello world!")
-	}
-
-	#[x]  // Specifying multiple cases indicates that the item would appear on all of them
-	#[y] .// This behaviour is designed to be different from `#[cfg(..)]`
-	fn a() -> i32 {
-		#[x]  // This statement would only appear on case `x`
-		let p = 1;
-		#[y]  // This statement would only appear on case `y`
-		let p = 2;
-		p + 1
-	}
-
-
-	fn g<#[x] 'a, #[y] T>(#[x] a: i32, #[y] a: u32) {}
+#[macro_export]
+macro_rules! if_production_or {
+	($prod_variant:expr, $non_prod_variant:expr) => {
+		if cfg!(feature = "production") {
+			$prod_variant
+		} else {
+			$non_prod_variant
+		}
+	};
 }
 
-```
-Expands to:
-```
-mod x {
-	fn a() -> i32 {
-		let p = 1;
-		p + 1
-	}
-	fn u() {
-		println!("hello world!");
-	}
-	fn g<'a>(a: i32) {}
-}
-
-mod y {
-	fn a() -> i32 {
-		let p = 2;
-		p + 1
-	}
-	fn v(a: String) {
-		println!("hello world!");
-	}
-	fn g<T>(a: u32) {}
-}
-
-```
-*/
-#[proc_macro_attribute]
-pub fn reuse(args: TokenStream, input: TokenStream) -> TokenStream {
-	handle_reuse(parse_macro_input!(args), parse_macro_input!(input))
-		.unwrap_or_else(Error::into_compile_error)
-		.into()
+#[macro_export]
+macro_rules! if_not_production {
+	($expression:expr) => {
+		if cfg!(not(feature = "production")) {
+			$expression
+		}
+	};
 }
