+++ primitives/core/src/lib.rs
@@ -17,7 +17,7 @@
 #![cfg_attr(not(feature = "std"), no_std)]
 #![allow(clippy::identity_op)]
 
-pub use litentry_macros::*;
+pub use litentry_proc_macros::*;
 
 pub use constants::*;
 pub use opaque::*;
