+++ tee-worker/app-libs/parentchain-interface/Cargo.toml
@@ -29,6 +29,7 @@ sp-runtime = { default-features = false, git = "https://github.com/paritytech/su
 
 # litentry
 lc-scheduled-enclave = { path = "../../litentry/core/scheduled-enclave", default-features = false, optional = true }
+litentry-hex-utils = { path = "../../../primitives/hex", default-features = false }
 litentry-primitives = { path = "../../litentry/primitives", default-features = false }
 sp-std = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.42" }
