+++ tee-worker/app-libs/stf/Cargo.toml
@@ -41,6 +41,7 @@ sp-std = { default-features = false, git = "https://github.com/paritytech/substr
 # litentry
 itp-node-api-metadata-provider = { path = "../../core-primitives/node-api/metadata-provider", default-features = false }
 lc-stf-task-sender = { path = "../../litentry/core/stf-task/sender", default-features = false }
+litentry-macros = { path = "../../../primitives/core/macros", default-features = false }
 litentry-primitives = { path = "../../litentry/primitives", default-features = false }
 pallet-parentchain = { path = "../../../pallets/parentchain", default-features = false }
