+++ tee-worker/app-libs/stf/src/getter.rs
@@ -19,7 +19,8 @@ use codec::{Decode, Encode};
 use ita_sgx_runtime::{IdentityManagement, System};
 use itp_stf_interface::ExecuteGetter;
 use itp_stf_primitives::{traits::GetterAuthorization, types::KeyPair};
-use itp_utils::{if_production_or, stringify::account_id_to_string};
+use itp_utils::stringify::account_id_to_string;
+use litentry_macros::if_production_or;
 use litentry_primitives::{Identity, LitentryMultiSignature};
 use log::*;
 use sp_std::vec;
