+++ tee-worker/cli/Cargo.toml
@@ -53,6 +53,7 @@ lc-credentials = { path = "../litentry/core/credentials" }
 # litentry
 frame-metadata = "15.0.0"
 ita-sgx-runtime = { path = "../app-libs/sgx-runtime" }
+litentry-hex-utils = { path = "../../primitives/hex", default-features = false }
 litentry-primitives = { path = "../litentry/primitives" }
 my-node-runtime = { package = "rococo-parachain-runtime", path = "../../runtime/rococo" }
 pallet-teerex = { path = "../../pallets/teerex", default-features = false }
