+++ tee-worker/enclave-runtime/src/rpc/worker_api_direct.rs
@@ -45,13 +45,14 @@ use itp_top_pool_author::traits::AuthorApi;
 use itp_types::{
 	DirectRequestStatus, Index, MrEnclave, RsaRequest, ShardIdentifier, SidechainBlockNumber, H256,
 };
-use itp_utils::{if_not_production, FromHexPrefixed, ToHexPrefixed};
+use itp_utils::{FromHexPrefixed, ToHexPrefixed};
 use its_primitives::types::block::SignedBlock;
 use its_sidechain::rpc_handler::{
 	direct_top_pool_api, direct_top_pool_api::decode_shard_from_base58, import_block_api,
 };
 use jsonrpc_core::{serde_json::json, IoHandler, Params, Value};
 use lc_scheduled_enclave::{ScheduledEnclaveUpdater, GLOBAL_SCHEDULED_ENCLAVE};
+use litentry_macros::if_not_production;
 use litentry_primitives::DecryptableRequest;
 use log::debug;
 use sgx_crypto_helper::rsa3072::Rsa3072PubKey;
@@ -489,7 +490,7 @@ fn forward_dcap_quote_inner(params: Params) -> Result<OpaqueExtrinsic, String> {
 
 	let param = &hex_encoded_params.get(0).ok_or("Could not get first param")?;
 	let encoded_quote_to_forward: Vec<u8> =
-		itp_utils::hex::decode_hex(param).map_err(|e| format!("{:?}", e))?;
+		litentry_hex_utils::decode_hex(param).map_err(|e| format!("{:?}", e))?;
 
 	let url = String::new();
 	let ext = generate_dcap_ra_extrinsic_from_quote_internal(url, &encoded_quote_to_forward)
@@ -519,7 +520,7 @@ fn attesteer_forward_ias_attestation_report_inner(
 
 	let param = &hex_encoded_params.get(0).ok_or("Could not get first param")?;
 	let ias_attestation_report =
-		itp_utils::hex::decode_hex(param).map_err(|e| format!("{:?}", e))?;
+		litentry_hex_utils::decode_hex(param).map_err(|e| format!("{:?}", e))?;
 
 	let url = String::new();
 	let ext = generate_ias_ra_extrinsic_from_der_cert_internal(url, &ias_attestation_report)
