+++ tee-worker/litentry/core/assertion-build/src/nodereal/amount_holding/evm_amount_holding.rs
@@ -123,9 +123,9 @@ pub fn build(
 mod tests {
 	use super::*;
 	use itp_stf_primitives::types::ShardIdentifier;
-	use itp_utils::hex::decode_hex;
 	use lc_credentials::assertion_logic::{AssertionLogic, Op};
 	use lc_mock_server::run;
+	use litentry_hex_utils::decode_hex;
 
 	fn create_ton_token_assertion_logic() -> Box<AssertionLogic> {
 		Box::new(AssertionLogic::Item { src: "$token".into(), op: Op::Equal, dst: "TON".into() })
