+++ tee-worker/litentry/core/vc-issuance/lc-vc-task-receiver/Cargo.toml
@@ -44,6 +44,7 @@ lc-data-providers = { path = "../../data-providers", default-features = false }
 lc-stf-task-receiver = { path = "../../stf-task/receiver", default-features = false }
 lc-stf-task-sender = { path = "../../stf-task/sender", default-features = false }
 lc-vc-task-sender = { path = "../lc-vc-task-sender", default-features = false }
+litentry-macros = { path = "../../../../../primitives/core/macros", default-features = false }
 litentry-primitives = { path = "../../../primitives", default-features = false }
 pallet-identity-management-tee = { path = "../../../pallets/identity-management", default-features = false }
