--- tee-worker/litentry/macros/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-authors = ["Trust Computing GmbH <info@litentry.com>"]
-name = "litentry-macros"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-cargo_toml = "0.16.3"
-quote = "1.0.33"
-
-[lib]
-proc-macro = true
