+++ tee-worker/litentry/primitives/src/lib.rs
@@ -41,7 +41,7 @@ pub use validation_data::*;
 use bitcoin::sign_message::{signed_msg_hash, MessageSignature};
 use codec::{Decode, Encode, MaxEncodedLen};
 use itp_sgx_crypto::ShieldingCryptoDecrypt;
-use itp_utils::hex::hex_encode;
+use litentry_hex_utils::hex_encode;
 use log::error;
 pub use parentchain_primitives::{
 	all_bitcoin_web3networks, all_evm_web3networks, all_substrate_web3networks, all_web3networks,
