+++ tee-worker/service/Cargo.toml
@@ -74,6 +74,8 @@ ita-stf = { path = "../app-libs/stf", default-features = false }
 lc-data-providers = { path = "../litentry/core/data-providers" }
 lc-mock-server = { path = "../litentry/core/mock-server" }
 lc-stf-task-sender = { path = "../litentry/core/stf-task/sender", default-features = false }
+litentry-hex-utils = { path = "../../primitives/hex", default-features = false }
+litentry-macros = { path = "../../primitives/core/macros", default-features = false }
 litentry-primitives = { path = "../litentry/primitives" }
 my-node-runtime = { package = "rococo-parachain-runtime", path = "../../runtime/rococo" }
 sgx-verify = { path = "../../pallets/teerex/sgx-verify", default-features = false }
