+++ tee-worker/sidechain/consensus/aura/src/lib.rs
@@ -37,7 +37,6 @@ use itp_sgx_externalities::SgxExternalities;
 use itp_stf_state_handler::handle_state::HandleState;
 use itp_time_utils::duration_now;
 
-use itp_utils::hex::hex_encode;
 use its_block_verification::slot::slot_author;
 use its_consensus_common::{Environment, Error as ConsensusError, Proposer};
 use its_consensus_slots::{SimpleSlotWorker, Slot, SlotInfo};
@@ -47,6 +46,7 @@ use its_primitives::{
 };
 use its_validateer_fetch::ValidateerFetch;
 use lc_scheduled_enclave::ScheduledEnclaveUpdater;
+use litentry_hex_utils::hex_encode;
 use sp_core::ByteArray;
 use sp_runtime::{
 	app_crypto::{sp_core::H256, Pair},
